Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    double float a;
+    float a;
     scanf("%f",&a);
     printf("You entered: %.2f",a);
     return 0;
